src/automatic-vector-db-updation/requirements.txt
fastapi~=0.115.6
uvicorn~=0.33.0
numpy~=1.24.4
matplotlib~=3.7.5
requests~=2.32.3
apscheduler~=3.11.0
python-dotenv~=1.0.1
opencv-python~=4.10.0.84
python-multipart~=0.0.20
msgpack~=1.1.0
ujson~=5.10.0
pymilvus~=2.5
psycopg2-binary~=2.9.10
typing-extensions~=4.12.2
schedule~=1.2.2
